+++ CHANGELOG.rst
@@ -14,10 +14,20 @@ paths are considered internals and can change in minor and patch releases.
 v4.28.1 (2024-0?-??)
 --------------------
 
+Added
+^^^^^
+- `TypeAliasType` support added (`#480
+  <https://github.com/omni-us/jsonargparse/issues/480>`__).
+
 Fixed
 ^^^^^
--  Failure to process ``Annotated`` dataclass members, and inclusion of
-   non-init fields in `attrs` and Pydantic dataclass-like instantiation.
+- Attrs and Pydantic 2 dataclasses with non-init fields fail to instantiate
+  (`#480 <https://github.com/omni-us/jsonargparse/issues/480>`__).
+- Default values/factories for Pydantic 2 dataclasses with `Field` initializers
+  are not right (`#480 <https://github.com/omni-us/jsonargparse/issues/480>`__).
+- `Annotated` fields in dataclass-likes (eg FastAPI types) resolve incorrectly
+  (`#480 <https://github.com/omni-us/jsonargparse/issues/480>`__).
+
 
 v4.28.0 (2024-03-??)
 --------------------